Add files via upload
This paper was written for a Programming Languages class, in which we dove deeper into the mentality behind programming and programming languages, along with exploring some new, lesser-used languages. In this paper we were instructed to use rosettacode.org to compare different languages and do our own research to come to our own conclusions about the causes of language popularity.